Edited seesaw scene
Edited seesaw scene and created the scene structure and worked on implementing the seesaw.

The seesaw seems to mostly work for rigidbody, but the characterbody isn't working yet.

Just uploading what I've done so far.

Also messed with a sandbox level scene, but it should be separate from everything else.